+++ internal/engineresolver/errwrapper_test.go
@@ -1,4 +1,4 @@
-package sessionresolver
+package engineresolver
 
 import (
 	"errors"
