+++ internal/engineresolver/jsoncodec.go
@@ -1,4 +1,4 @@
-package sessionresolver
+package engineresolver
 
 //
 // JSON codec
