+++ internal/engineresolver/lookup.go
@@ -1,4 +1,4 @@
-package sessionresolver
+package engineresolver
 
 //
 // Actual lookup code
